#759| Changing is present to check for presence of variables except for 0. Handled for calculations

+++ modules/db/src/main/resources/db/migration/V40__create_program_product_isa.sql
@@ -4,11 +4,11 @@
 DROP TABLE IF EXISTS program_product_isa;
 CREATE TABLE program_product_isa (
     id SERIAL PRIMARY KEY,
-    whoRatio NUMERIC (8,4) NOT NULL,
+    whoRatio INTEGER NOT NULL,
     dosesPerYear INTEGER NOT NULL,
-    wastageRate NUMERIC(6,2) NOT NULL,
+    wastageRate INTEGER NOT NULL,
     programProductId INTEGER REFERENCES program_products(id),
-    bufferPercentage NUMERIC(6,2) NOT NULL,
+    bufferPercentage INTEGER NOT NULL,
     minimumValue INTEGER,
     adjustmentValue INTEGER NOT NULL,
     calculatedIsa INTEGER,
